feat(structure): extractor registry + JS/TS/ESM extractor in scripts/lang/ (T04)
update-structure.js now dispatches parsing through a per-language extractor
registry. The CommonJS logic moves verbatim into scripts/lang/javascript.js
(golden fixture stays byte-identical) and gains ESM support (import/export,
re-export, dynamic and side-effect imports), the .mjs/.cjs/.jsx/.ts/.tsx
extensions, TS param-annotation cleanup (applied only when a ":" annotation
is present, so JS output is unchanged), export-prefixed declarations and TS
generics. Module keys and --changed filtering derive from the registry's
extension set. structureBootstrap ships scripts/lang/* into user projects'
.frame/bin/lang/. STRUCTURE.json picks up the sample project's .jsx modules
— siblings of its already-indexed .js files, discovered by the broader
extension set.
