+++ src/main/structureBootstrap.js
@@ -86,6 +86,24 @@ function copyParserScripts(projectPath) {
       console.warn(`[frame] failed to copy ${file}: ${err.message}`);
     }
   }
+
+  // Ship the per-language extractors (scripts/lang/*) alongside the parser —
+  // update-structure.js requires them relative to its own location.
+  const langSrcDir = path.join(SCRIPTS_SOURCE_DIR, 'lang');
+  if (fs.existsSync(langSrcDir)) {
+    const langDstDir = path.join(binDir, 'lang');
+    if (!fs.existsSync(langDstDir)) {
+      fs.mkdirSync(langDstDir, { recursive: true });
+    }
+    for (const file of fs.readdirSync(langSrcDir).filter((f) => f.endsWith('.js'))) {
+      try {
+        fs.copyFileSync(path.join(langSrcDir, file), path.join(langDstDir, file));
+        copied.push(`lang/${file}`);
+      } catch (err) {
+        console.warn(`[frame] failed to copy lang/${file}: ${err.message}`);
+      }
+    }
+  }
   return copied;
 }
